+++ firmware/controllers/algo/accel_enrichment.cpp
@@ -71,19 +71,6 @@ float TpsAccelEnrichment::getTpsEnrichment() {
 		resetFractionValues();
 	}
 
-#if EFI_TUNER_STUDIO
-	if (engineConfiguration->debugMode == DBG_TPS_ACCEL) {
-		engine->outputChannels.debugFloatField1 = tpsFrom;
-		engine->outputChannels.debugFloatField2 = tpsTo;
-		engine->outputChannels.debugFloatField3 = valueFromTable;
-		engine->outputChannels.debugFloatField4 = extraFuel;
-		engine->outputChannels.debugFloatField5 = accumulatedValue;
-		engine->outputChannels.debugFloatField6 = maxExtraPerPeriod;
-		engine->outputChannels.debugFloatField7 = maxInjectedPerPeriod;
-		engine->outputChannels.debugIntField1 = cycleCnt;
-	}
-#endif /* EFI_TUNER_STUDIO */
-
 	float mult = interpolate2d(rpm, config->tpsTspCorrValuesBins,
 						config->tpsTspCorrValues);
 	if (mult != 0 && (mult < 0.01 || mult > 100)) {
@@ -194,29 +181,21 @@ TpsAccelEnrichment::TpsAccelEnrichment() {
 	cb.setSize(4);
 }
 
-#if ! EFI_UNIT_TEST
+void TpsAccelEnrichment::onConfigurationChange(engine_configuration_s const* /*previousConfig*/) {
+	constexpr float slowCallbackPeriodSecond = SLOW_CALLBACK_PERIOD_MS / 1000.0f;
+	int length = engineConfiguration->tpsAccelLookback / slowCallbackPeriodSecond;
 
-static void setTpsAccelLen(int length) {
 	if (length < 1) {
 		efiPrintf("setTpsAccelLen: Length should be positive [%d]", length);
 		return;
 	}
-	engine->tpsAccelEnrichment.setLength(length);
+
+	setLength(length);
 }
-
-void updateAccelParameters() {
-	constexpr float slowCallbackPeriodSecond = SLOW_CALLBACK_PERIOD_MS / 1000.0f;
-	setTpsAccelLen(engineConfiguration->tpsAccelLookback / slowCallbackPeriodSecond);
-}
-
-#endif /* ! EFI_UNIT_TEST */
-
 
 void initAccelEnrichment() {
 	tpsTpsMap.initTable(config->tpsTpsAccelTable, config->tpsTpsAccelToRpmBins, config->tpsTpsAccelFromRpmBins);
 
-#if ! EFI_UNIT_TEST
-	updateAccelParameters();
-#endif /* ! EFI_UNIT_TEST */
+	engine->module<TpsAccelEnrichment>()->onConfigurationChange(nullptr);
 }