+++ firmware/controllers/algo/fuel_math.cpp
@@ -328,7 +328,7 @@ float getInjectionMass(int rpm) {
 		engine->module<InjectorModelSecondary>()->prepare();
 	}
 
-	float tpsAccelEnrich = engine->tpsAccelEnrichment.getTpsEnrichment();
+	float tpsAccelEnrich = engine->module<TpsAccelEnrichment>()->getTpsEnrichment();
 	efiAssert(ObdCode::CUSTOM_ERR_ASSERT, !std::isnan(tpsAccelEnrich), "NaN tpsAccelEnrich", 0);
 	engine->engineState.tpsAccelEnrich = tpsAccelEnrich;
 
@@ -444,7 +444,7 @@ float getStandardAirCharge() {
 	return idealGasLaw(cylDisplacement, 101.325f, 273.15f + 20.0f);
 }
 
-float getCylinderFuelTrim(size_t cylinderNumber, int rpm, float fuelLoad) {
+PUBLIC_API_WEAK_SOMETHING_WEIRD float getCylinderFuelTrim(size_t cylinderNumber, float rpm, float fuelLoad) {
 	auto trimPercent = interpolate3d(
 		config->fuelTrims[cylinderNumber].table,
 		config->fuelTrimLoadBins, fuelLoad,