+++ firmware/CHANGELOG.md
@@ -44,6 +44,8 @@ Release template (copy/paste this for new release):
  - STM32Cube v2.17
  - Better validation of thermistor configuration #6724
  - Full layout vs tuning layout #5154
+ - Usability: turn on boost controller without ECU power cycle #6761
+ - Knock frequency inputs to be bore diameter and checkbox #6767
 
 ### Fixed
  - knock logic not activated until any configuration change via TS #6462